vold: add getPartDevice() method to Volume
Formatting and sharing currently operate on the disk device, even
when a partition is specified in vold.fstab. In this case these
operations should act only on the partition to avoid data loss or
corruption for other device partitions. The getPartDevice() method
is added to the Volume class for this purpose, and the formatting
and sharing code is updated to use this new method.

Signed-off-by: Seth Forshee <seth.forshee@garmin.com>