Unmount file system on incorrectly removed card
Unmount any mounted file system on a not partitioned card if that
card is abruptly removed. Previously only file systems mounted on
partitioned cards where unmounted if the card was removed abruptly.

This patch adds a check when handling the disk removed event if a
file system is mounted on the disk then it is unmounted.

Change-Id: I18843aeb340ee34d90f80e413b135a5d33ac28e9
Signed-off-by: Benn Porscke <benn.porscke@stericsson.com>