Unmount file system on incorrectly removed card
Unmount any mounted file system on a not partitioned card if that
card is abruptly removed. Previously only file systems mounted on
partitioned cards where unmounted if the card was removed abruptly.

This patch adds a check when handling the disk removed event if a
file system is mounted on the disk then it is unmounted.

Change-Id: Id1f66bdea85737e946e01d83f6aa2917f09bb0f7
Signed-off-by: Mikael Gullberg <mikael.xx.gullberg@stericsson.com>
Signed-off-by: Christian Bejram <christian.bejram@stericsson.com>