Refine the return value check of fat fsck
When the volume checking each partitions for avaliable
VFAT filesystem, we will get ENOEXEC return value on
exec msdos_fsck who give out "Invalid cluster size: 0"
on the extension partition.
When this error returns, it should not block volume manager
to scan the left partitions as it's not a I/O error.
This issue happend when we have a extension partition
on the internal or external storage.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>